Add files via upload
Both methods; including the "efficient" method, take ages. I'm not sure if its because ruby is not good at these types of calculations or because im using a laptop with a lot of additional processes going on in the background. I have heard that ruby struggles with larger calculations like this.